Add support for (Async)Connection
Added support for (Async)Connection when initializing (Async)AzurePGVectorStore.
When a connection pool is given, the vector store will get a connection from the
pool. Otherwise, it will use the given connection.

This feature is added to keep the symmetry between langchain-azure-postgresql
and langgraph-checkpoint-postgres. In other words, because
`langgraph-checkpoint-postgres` supports (Async)Connection, it is useful to have
the same feature in `langchain-azure-postgresql`.

Note that, for the changes to take effect, users *must* pass the connection
object with autocommit set to True. Otherwise, the changes may not be applied as
expected.